Rewrite %tconvert to properly handle minutes
Pre-existing bug existed in code where minutes in the time or timezone
would not be handled properly and result in a wrong time result.
Whenever the difference exceeded 60 min, it would roll over minutes
properly but would not subtract an hour from the hours. Instead, it now
handles everything by first converting to minutes, then back to
hour/minute format.